requirements.txt
face_recognition
opencv-python
pyobjc-framework-Quartz
numpy
keyring